ex00 finished

+++ Module_00/ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: alaaouam <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/19 13:14:38 by alaaouam          #+#    #+#             */
-/*   Updated: 2023/07/19 14:34:11 by alaaouam         ###   ########.fr       */
+/*   Updated: 2023/07/19 15:40:41 by alaaouam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int	main(int argc, char **argv)
 		while (count < argc)
 		{
 			pos = 0;
-			while (pos < strlen(argv[count]))
+			while (pos < (int)strlen(argv[count]))
 			{
 				std::cout << (char)toupper(argv[count][pos]);
 				pos++;
